Rearrange packages to avoid circular dependencies
